k8s: move init test to new watcher_test.go
This commit extracts the k8sWatcher related unit test into it's own
file `watcher_test.go`. (Separate commit to keep the git history).
